requirements.txt
asgiref==3.2.3
boto3==1.11.5
botocore==1.14.5
certifi==2019.11.28
Collectfast==1.3.1
dj-database-url==0.5.0
Django==3.0.1
django-crispy-forms==1.8.1
django-storages==1.8
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Pillow==7.0.0
pipenv==2018.11.26
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.1
six==1.14.0
-e git://github.com/jazzband/sorl-thumbnail@4fe1854ee30e24adfc57884342517bbf506d0cd9#egg=sorl_thumbnail
sqlparse==0.3.0
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
whitenoise==5.0.1
